% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{AnalyzeOutput_mLoci_Daisy}
\alias{AnalyzeOutput_mLoci_Daisy}
\title{Analyze output for mPlex-mLoci or DaisyDrive}
\usage{
AnalyzeOutput_mLoci_Daisy(readDirectory, saveDirectory = NULL, filename,
  genotypes, collapse)
}
\arguments{
\item{readDirectory}{directory where output was written to; should not end in path seperator}

\item{saveDirectory}{directory to save analyzed data. Default is readDirectory}

\item{genotypes}{A list of each locus containing the genotypes of interest at that locus. Default is all genotypes}

\item{collapse}{A vector of each locus containing TRUE/FALSE. If TRUE, the genotypes of interest at that locus are collapsed and the output is the sum of all of them.}
}
\value{
A *.rds object containing list(metaData=character, maleData=array, femaleData=array)
}
\description{
This function takes all the files in a directory and analyzes the population by
genotype of interest. It saves output by run, and inside each run it contains
arrays corresponding to the male and female of each patch. The arrays are organzed
by time, then each genotype and total population, and the array depth is each
patch. Data are analyzed by matching the genotypes of interest.
}
