% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{CreateMosquitoes_Distribution_Genotype}
\alias{CreateMosquitoes_Distribution_Genotype}
\title{Create new Mosquito objects}
\usage{
CreateMosquitoes_Distribution_Genotype(numMos, minAge, maxAge, ageDist, aTypes)
}
\arguments{
\item{numMos}{Integer number of Mosquitoes to create}

\item{minAge}{Integer specifying the minimum age of Mosquitoes}

\item{maxAge}{Integer specifying the maximum age of Mosquiotes}

\item{ageDist}{Distribution for ages of Mosquitoes. Must be length(maxAge-minAge+1)}

\item{aTypes}{Nested list containing the alleles and their probabilities at each locus}
}
\value{
List of Mosquito objects length(sum(numMos))
}
\description{
Creates new Mosquito objects specifically for simulation setup
}
\details{
This function creates new mosquitoes. It is similar to
\code{\link{CreateMosquitoes_Defined_Genotype}}, but intended for simulation setup.
It takes a list of alleles and their frequency in the population, then creates
genotypes based on that distribution. Then it samples the age distribtuion, and
creates new Mosquitoes with the proper distributions of genotypes and age.
}
\examples{
set.seed(42)
numMos <- 10
minAge <- 1
maxAge <- 10
ageDist <- rep(x = 1, times = 10-1+1)/(10-1+1) #uniform distribution

alleloTypes <- vector(mode = "list", length = 2) #2 loci
alleloTypes[[1]]$alleles <- c("W","H")
alleloTypes[[1]]$probs <- c(1,0)
alleloTypes[[2]]$alleles <- c("W","H")
alleloTypes[[2]]$probs <- c(1,0)

CreateMosquitoes_Distribution_Genotype(numMos, minAge, maxAge, ageDist, AlleloTypes)

}
