% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{JaredDirichlet}
\alias{JaredDirichlet}
\title{R Dirichlet}
\usage{
JaredDirichlet(n, alpha)
}
\arguments{
\item{n}{Number of draws to perform}

\item{alpha}{vector length(n) as a shape parameter}
}
\value{
matrix(nrow=n, ncol=length(alpha))
}
\description{
Calculates a Dirichlet distribution.
}
\details{
This function makes draws from a Dirichlet dristribution. It is slow,
but entirely written in R.
}
\examples{
set.seed(42)
JaredDirichlet(n=4, alpha = c(0.1,0.2,0.3,0.4))

}
