% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_MultiplexGeneratingFunction_multiLocus.R
\name{MultiplexOffspring_mLoci}
\alias{MultiplexOffspring_mLoci}
\title{mPlex Multiple Loci Offspring}
\usage{
MultiplexOffspring_mLoci(fGen, mGen, reference)
}
\arguments{
\item{fGen}{Female genotype}

\item{mGen}{Male genotype}

\item{reference}{Offspring reference list}
}
\value{
List(Alleles, Probabilities)
}
\description{
Create a list of offspring genotypes and their probabilities
}
\details{
Using the reference generated by \code{\link{MakeReference_Multiplex_mLoci}},
this function expands the possible genotypes of the offspring and the ratios
that they occur. Similar to \code{\link{DaisyOffspring}} and
\code{\link{MultiplexOffspring_oLocus}}.
}
\examples{
ref <- MakeReference_Multiplex_mLoci(H = 0.98, R = 0.001, S = 0.0003, d = .00001)
fGen <- "WW"
mGen <- "WW"

MultiplexOffspring_mLoci(fGen, mGen, ref)

}
