% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_LarvaMature_Patch}
\alias{oneDay_LarvaMature_Patch}
\title{Daily Larva Maturation}
\usage{
oneDay_LarvaMature_Patch()
}
\description{
Larva mature into pupa. The average time spent as a larva is specified in
stageTime("L"). This is used to parametrize a lognormal function with
standard deviation log(1.2), slightly right-skewing the distribution.
}
