% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchClass.R
\name{oneDay_migrationOut_Patch}
\alias{oneDay_migrationOut_Patch}
\title{Oubound Migration}
\usage{
oneDay_migrationOut_Patch()
}
\description{
Stochastic model of migration of adults from this patch.
}
\details{
Migration is modeled as a Dirichlet-Multinomial process parameterized
by \code{moveVar} multiplied by the row corresponding to this patch from the
migration matrix. A Dirichlet distributed random variate is sampled over the
other patch frequency and then from \code{JaredDirichlet} according to that
parameter vector and then movement is sampled from \code{\link[stats]{rmultinom}}.
}
