% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchClass.R
\name{oneDay_Releases_Patch}
\alias{oneDay_Releases_Patch}
\title{Release Mosquitoes in a Patch}
\usage{
oneDay_Releases_Patch()
}
\description{
Based on this patch's release schedule, handle daily releases.
}
