% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{Plot_mPlex}
\alias{Plot_mPlex}
\title{Plot mPlex}
\usage{
Plot_mPlex(file, totalPop=TRUE)
}
\arguments{
\item{file}{Path to .gzip file from analyze function}

\item{totalPop}{Boolean, to plot the total population or not.}
}
\description{
Plots the analyzed output of mPlex.
}
\details{
This function plots output from the analyze function. Setting totalPop
to FALSE keeps it from plotting the total population.
}
