% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{splitOutput}
\alias{splitOutput}
\title{Split Output by Patch}
\usage{
splitOutput(directory)
}
\arguments{
\item{directory}{directory where output was written to; must not end in path seperator}

\item{multiCore}{Set whether or not to use multiple cores for writing output}
}
\value{
*.csv files for each patch and each run
}
\description{
Split each run output into multiple files by patch.
}
